MeshML-master/services/api-gateway/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# gRPC
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin bcrypt version for compatibility with passlib
python-multipart==0.0.9

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# GCS
google-cloud-storage==2.16.0
boto3==1.34.51

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
orjson==3.9.15
tenacity==8.2.3
pandas==2.2.0
openpyxl==3.1.2

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.2.0
ruff==0.2.2
mypy==1.8.0
pre-commit==3.6.2
